Refactoring 'utils'
