+++ 2023/internal/utils/reversearray.go
@@ -1,0 +1,8 @@
+package utils
+
+func ReverseArray[T any](a []T) []T {
+	for i, j := 0, len(a)-1; i < j; i, j = i+1, j-1 {
+		a[i], a[j] = a[j], a[i]
+	}
+	return a
+}
